Basic materials support.
The only property the current fragment shader appears to support is the
color from a diffuse texture (okay, well really ambient, but they're the
same file in all these materials, and I suspect the lighting will switch
to diffuse-only before it goes full Blinn-Phong), so that's the only
property supported.

Currently the OBJ loader as a whole leaks everything it returns to C++.
(And it's *not* safe to free() or delete or delete[] parts of it from
C++!) However, it's structured so that in theory, a free_obj() function
could be written to take a ModelsAndMaterials and free the backing
memory pretty easily. (The only ABI-breaking change that should result
would be changing the vector-structs to include a capacity field.)